💶Update data types for decimals, adult, and child in Token and IBukProtocol to optimize gas⛽️

+++ contracts/BukProtocol/IBukProtocol.sol
@@ -33,8 +33,8 @@ interface IBukProtocol {
      * @param uint256 tokenId           Token ID.
      * @param bytes32 propertyId        Property ID.
      * @param BookingStatus status      Booking status.
-     * @param uint8 adult[]             Number of adults.
-     * @param uint8 child[]             Number of children.
+     * @param uint256 adult[]             Number of adults.
+     * @param uint256 child[]             Number of children.
      * @param address owner             Address of the booking owner.
      * @param uint256 checkin           Check-in date.
      * @param uint256 checkout          Check-out date.
@@ -49,8 +49,8 @@ interface IBukProtocol {
         uint256 tokenId;
         bytes32 propertyId;
         BookingStatus status;
-        uint8 adult;
-        uint8 child;
+        uint256 adult;
+        uint256 child;
         address firstOwner;
         uint256 checkin;
         uint256 checkout;
@@ -113,8 +113,8 @@ interface IBukProtocol {
         bytes32 indexed propertyId,
         uint256 checkin,
         uint256 checkout,
-        uint8 adult,
-        uint8 child
+        uint256 adult,
+        uint256 child
     );
 
     /// @dev Emitted when booking refund is done.
@@ -203,7 +203,7 @@ interface IBukProtocol {
      * @param _commission Commission percentage.
      * @notice This function can only be called by admin
      */
-    function setCommission(uint8 _commission) external;
+    function setCommission(uint256 _commission) external;
 
     /**
      * @dev Function to toggle the tradeability of an asset.
@@ -242,8 +242,8 @@ interface IBukProtocol {
         uint256[] memory _total,
         uint256[] memory _baseRate,
         uint256[] memory _minSalePrice,
-        uint8[] memory _adult,
-        uint8[] memory _child,
+        uint256[] memory _adult,
+        uint256[] memory _child,
         bytes32 _propertyId,
         uint256 _checkin,
         uint256 _checkout,
